CQL: adding railroad diagram for the json output to the documentation
Summary: Note that the grammar is stricter than the JSON form. For instance JSON allows re-ordering of named items.  You shouldn't assume the order will never change.  However, the checked in grammar is used to verify the JSON output so you can be sure that the order in the grammar is the order that the compiler is currently producing.  The diagram might be stale but an authoritative one can be regenerated easily enough from the `grammar_docs` directory.

Reviewed By: RaoulFoaleng

Differential Revision: D24208644

fbshipit-source-id: 37e6c1d3eaa09b48ebd3ee6e1e0598e92e657a73