Adding trim, rtrim, and ltrim because they are popping up all over
Summary:
Semantic analysis for the trim trio is the same.  Only supported in the SQL context (though these
are actually doable in other contexts) 1 or 2 args, both args are strings, the result type is
the same as the type of the first arg.  But if either string is sensitive the result is sensitive.

Reviewed By: RaoulFoaleng

Differential Revision: D24429826

fbshipit-source-id: 1328080ba06618e7735cab962674bf8371d9c3cc