Add semantic analysis for length builtin func
Summary:
Add semantic analysis for length() builtin func

```
length(X)

For a string value X, the length(X) function returns the number of characters (not bytes) in X prior to the first NUL character. Since SQLite strings do not normally contain NUL characters, the length(X) function will usually return the total number of characters in the string X. For a blob value X, length(X) returns the number of bytes in the blob. If X is NULL then length(X) is NULL. If X is numeric then length(X) returns the length of a string representation of X.
```

Reviewed By: ricomariani

Differential Revision: D24439610

fbshipit-source-id: 9fd14665eda657b6feba3d089999b9738ec9d221