Allow creation of cursor shapes using procedure arguments (new) in addition to procedure results (already supported)
Summary:
The idea here is that you can use the types of a procedures arguments to get a useful shape.  This lets you do some
very interesting like

```
create proc foo_shim(like foo arguments)
begin
   call log_stuff(foo_arg_1);
   call foo(from arguments);
end;
```
or

```
create proc test_something()
begin
   declare C cursor like something arguments;
   fetch args from values (1, 3, 4, 5);
   call something(from C);

   update cursor C(x) values(3);
   call something(from C);

   etc.
end;
```
coming soon:

```
fetch args using  1  x,
                  2  y,
                  3  z;
```
which when combined with the above gives you a clear and type-safe and not sensitive to ordering way to call a proc with lots of arguments.  This comes up all the time in test code.

Reviewed By: RaoulFoaleng

Differential Revision: D24747318

fbshipit-source-id: 0568969a099243d77600debd74f90e8fef0982e6