CQL: add @attribute(cql:suppress_result_set) and increase stack size
Summary:
The scratch variables stack only worked up to 64 levels but it's easily generalized.  64 seemed like a lot of levels but recent rewrite features can easily blow past this so it's now doubled to 128.  It's very easy to make it large at this point.

Additionally there are lots of cases where you don't need the result set features for a procedure that returns one.  You might be just making a helper that boxes or otherwise produces a cursor.  You can use `attribute(cql:suppress_result_set)` to suppress that now.

Reviewed By: RaoulFoaleng

Differential Revision: D24800545

fbshipit-source-id: 87eddbbe9e86615796e67f35c3555a78a64a9c58