Add built in func iif to cql
Summary:
Add built in func `iif` to cql.

The `iif(X,Y,Z)` function returns the value `Y` if `X` is true, and `Z` otherwise. The `iif(X,Y,Z)` function is logically equivalent to and generates the same bytecode as the `CASE` expression `CASE WHEN X THEN Y ELSE Z END`

In CQL it is literally rewritten to be the aforementioned CASE expression.

Reviewed By: ricomariani

Differential Revision: D24892777

fbshipit-source-id: c18b94c1e13a94e4413c8794123ea3e795be597b