Enable tracing of CQL errors with a custom macro; throw doesn't lose the error code
Summary:
Enable tracing of CQL errors with a custom macro cql_error_trace() which you redefine as needed in your cqlrt

This diff also improves the `throw` code gen so that a rethrow will definitely throw the error code that came in, even if there are intervening successes.

`cql_error_trace` can be used to harvest the `sqlite3_errmsg` immediately.  This can be problematic otherwise because there are often intervening successes due to cleanup logic after the procedure runs.

Note that this is quite expensive so you don't want to be doing this all the time.

Reviewed By: RaoulFoaleng

Differential Revision: D24995565

fbshipit-source-id: f314adedb50dd473adf6b2fd5e9b8795d76caa20