LIKE expression escape doesn't mean what we think it means
Summary:
Reported by mingodad at https://github.com/facebookincubator/CG-SQL/issues/19

This isn't a real problem for us because that code only runs on old sqlite versions which don't have the new
postgress compatible meaning of `.. like .. escape`.  However we want the tests to pass in compat mode even
on new SQLite and we really should use the escape parameter correctly.  The default is *no* escape not
'%' escape.  Adjust accordingly so we don't have to redo this when we support LIKE ESCAPE and so that the tests pass if you use sqlite trunk.

Reviewed By: toddkrabach

Differential Revision: D25065622

fbshipit-source-id: 34c64ef9c01623fea1e9332b3e70a4210127e0aa