Coverage additions from GCC coverage on Ubuntu (#21)
Summary:
There were a number of missing lines of coverage in cql.y that were discovered by gcc on Ubuntu.  This led to examining the WINDOW family of functions and adding test cases to parse them and generate their output.  There were several bugs in this code mostly to do with handling of spacing but a few where the AST was wrong.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/21

Test Plan: cov.sh

Reviewed By: RaoulFoaleng

Differential Revision: D25204497

Pulled By: ricomariani

fbshipit-source-id: 5f312f9699efcec49bdd1f8fdfc4edbcab8b32dc