Adding support for "x like shape, y like shape" in procedure arguments (#23)
Summary:
The idea here is that you you might want to create a procedure that takes two or more of something as an argument.  For instance:

  create table Person (
    id integer not null,
    name text not null
  );

  create procedure foo(p1 like Person, p2 like Person)
  begin
    ..
  end;

This procedure would have arguments:

  p1_id, p1_name, p2_id, p2_name

Which could then be used maybe like this

  call insert_person(from p1);
  call insert_person(from p2);

the arguments can be accessed by name `p1_name` etc.  or equivalently `p1.name`

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/23

Test Plan: cov.sh

Reviewed By: RaoulFoaleng

Differential Revision: D25205288

Pulled By: ricomariani

fbshipit-source-id: 409b5882d8807d0ef9c2fc4fd10dcfafcc5e651a