const(...) function added to give access to constant folding (#22)
Summary:
In order to do a good job with enumeration support we need constant numeric expressions.  This is the foundational work for that effort.  The idea here is to enable expression evaluation for numeric types by writing:

  select const(1+3);

which is evaluated at compile time resulting in

  select 4;

Of course this is really only helpful if there are macros and such involved, or, as we'll have later, named enumerations.

The constant folding is done by a simple ast rewrite of the call node.

Testing this also revealed a bug in the IIF rewrite:

```
sqlite> select iif(1, 'the truth', 'a lie');
the truth
sqlite> select iif(0, 'the truth', 'a lie');
a lie
sqlite>
```
We had args 2 & 3 exchanged.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/22

Test Plan:
test.sh
cov.sh

Reviewed By: RaoulFoaleng

Differential Revision: D25205307

Pulled By: ricomariani

fbshipit-source-id: 25aa887c5f64165e5aaa268bba77a8114587e378