Minor cleanups in arg bundle handling (#25)
Summary:
The declare proc statement handling did not set up the arg bundle name table so it couldn't use arg bundles.  It doesn't do anything with the bundles so this is just a case of make the table then throw it away so it's not null.

The term "cursor args" was used in many places where it could be a cursor or an arg_bundle.  This is now the more general term shape args.  There's no code change here it's just renames for clarity.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/25

Test Plan: cov.sh

Reviewed By: RaoulFoaleng

Differential Revision: D25294159

Pulled By: ricomariani

fbshipit-source-id: 6e153d6772d7beca46317875574b57bbb0b4b19c