CQL: Upgrade the const function to a rewriting primitive
Summary:
We finalize the meaning of "const" you can now use const in any basic expression context
and the places that only accepted a numeric literal before now also accept a
const expression hence:

  create table foo
  (
    id integer primary key,
    age integer default const(12+13)
  );

which is a bogus example but the point is const works where you could otherwise only
put literals.  This makes macros with const expressions much more useful.

Testing this revealed we still had some cases where we might emit the sequence "--" into SQL
e.g. `select - - 1;`  this was fixed by forcing a space if needed.  The C generation did not have this problem.

Testing also revealed that the data type of the default value for column definitions was not checked for compatibility against the column type. This was also corrected.

Reviewed By: RaoulFoaleng

Differential Revision: D25322971

fbshipit-source-id: f05fbdc7d2c7a53f5950bf6e2f4d53356913c6a8