compute arg info and the synthentic "arguments" shape
Summary:
In order to be able to emit the original information for argument lists we record the incoming data in arg_info.

We need this so that we can emit JSON that tells people how the arg list was made (using like X or whatever)

At this point we add a synthetic arg_bundle ARGUMENTS with this data.  But it is not yet used.

Reviewed By: RaoulFoaleng

Differential Revision: D25378977

fbshipit-source-id: d6a73dd24100188629e8ce20c13ad85f5400247f