JSON output for arg bundle types and other sources of complex args
Summary:
The json for args gets an "argOrigin" string which describes how that argument came to be
this is only interesting if the argument came from a LIKE construct.

There were a couple of places where the shape type info were not saved well (esp the original name) in the AST.
I had to fix those to get meaningful JSON results.

Reviewed By: RaoulFoaleng

Differential Revision: D25401138

fbshipit-source-id: 1aa256df1ab48fb00c6762e66a3a98ebcb764fa7