Conslidate from_cursor -> from_shape
Summary:
The AST node from_cursor can be a shape anywhere so it's been renamed to from_shape.

In the rewrite code there were many references to 'cursor' that are actually any shape.

Fixed a few comments and renamed a few local variables as well.

This is a no-op change.

Reviewed By: wanaya, RaoulFoaleng

Differential Revision: D25429414

fbshipit-source-id: c8353b7caac93f5c9a859f9d10d7dfd5dd0267c4