Add IDE debug support for CQL (#29)
Summary:
The key thing that's missing to be able to debug directly into the CQL in an IDE is the #line directives.

This is the first installment of such directives for top level statements only.  Now that means that if you're
trying to debug into say a CASE expression it's not going to help you very much so there's still quite a bit
of work here before this is finished but even with just this much support you can do a lot of interactive debugging.

Clicking on lines to set breakpoints in the IDE doesn't seem to work yet though you can set line breakpoints
with the lldb commands:

```
b  foo.sql:22
```

Symbolic breakpoints work fine as does single stepping.

NOTE: I figured out what's going on there, the IDE has to be configured to think that a .sql file can have breakpoints in it.  I'm looking into how to do that now.  If you rename the .sql file to .c and compile it with CQL it works... which I can't recommend but it means nothing is wrong with the debug info.  You can fix this by adding { "language" : "cql" } to the breakpoints section of package.json for your debug extension.  You may have to declare that language as well if you don't have an extension that already does so.

We also get the new `--nolines` option to suppress the `#` directives for lines (i.e. the old mode).  This is useful if you need to debug the C.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/29

Reviewed By: RaoulFoaleng

Differential Revision: D25465100

Pulled By: ricomariani

fbshipit-source-id: dbe337dc100dc57e22c1ede4c3bce2d02b77244a