Adding test tool to verify line number emission and a reference baseline
Summary:
The idea here is to parse the output C looking for line directives for any give proc.  We do this for the expected line numbers
and the actual line numbers.  A small test helper written in CQL stores the data for the lines of input and output.

To make this more resilient all the line numbers are changed so that they are relative to the start of the procedure.  This means
the expected line numbers do not change if you add more test cases.

The difference output flags the first line or text that varied between the reference and the result.  This also can be used to do
significant strict codegen tests.

Reviewed By: RaoulFoaleng

Differential Revision: D25501539

fbshipit-source-id: f4e29099baf37b5c6d2aa620ac1870c6a1a59d0a