Rename cursor variables for superior debugging (#33)
Summary:
in:
```
  declare C cursor for select 1 x, 2 y;
  fetch C
  if C.x == 1 then
    ...
  end if;
```

The generated code had the variable `C` representing the statement and the variable `C_` representing the loaded values.  This meant that you had to do  `p C_.x` to see what's in the cursor.  This change makes it so that the SQLite statement is stored in a variable named `C_stmt` and the variable `C` refers to the data.  Hence `p C.x` just works.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/33

Test Plan: ./cov.sh

Reviewed By: RaoulFoaleng

Differential Revision: D25519034

Pulled By: ricomariani

fbshipit-source-id: 19d0f5799ea11b99877b8cfd88bab6c3923b63ca