Emit enums stmt
Summary:
When we added enums to the language the code emitted them all into every .h file.  That's a bit crazy because as enums grow you'll just get these enormous .h files with every enum going into every header.

This change makes it so that the .h file gets the enums only if you ask for them.  The idea is that you can choose one .sql file put `emit_enums foo, bar` in there and then you'll get those enums in its header file.  That way you can make a nice logical place for your enums to live.

`emit_enums;` emits all enums that are visible in the translation unit (even those that came after).

NOTE: declared long enums were using `L` for the long values but on 32 bit systems it has to be `LL`.  So we use the `_64` macro like the other cases of this.  Noticed this in testing.

Reviewed By: RaoulFoaleng

Differential Revision: D25619262

fbshipit-source-id: 91e4fc7b8e94db3d4c983336d0a97dcd8188c6e2