in compressed CQL output, One trailing space in the fragment would cause it to go off the end.
Summary:
In compressed CQL output, One trailing  space in the fragment would cause it to go off the end.

There actually are no such fragments as the canonicalizer removes trailing space but we created one with buggy virtual table output accidentally which caused this failure.

Virtual tables no longer create the trailing
space but the code fix is in place and a special unit test to force the issue was added

Differential Revision: D25648587

fbshipit-source-id: cfe17a72791f4d96dbb296e82a2cc61f3c707e8b