USES TRANSACTION was not properly emitted for OUT UNION procs in the exports file
Summary:
This is just a missing case, USES TRANSACTION is necessary to specify that the proc uses the database and should therefore get the db parameter.
OUT and OUT UNION procs may or may not use the database to get their result so you have to do this explicitly

To avoid this kind of problem the exports are now emitted into the test stream where they can be verified.  I added test cases for all the flavors.

Reviewed By: RaoulFoaleng

Differential Revision: D25827117

fbshipit-source-id: a12642986ce26a8e4ece01029437bb7ca26e6ef8