Simplify named type processing a little bit and cleaned up comments (#40)
Summary:
This is a no-op change.  It's just minor cleanups.  Note we don't need the find_data_type_of_named_type function because the sem info is always on the node anyway.  We don't need to go to the ast that is the data type in the decl and we don't want to.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/40

Test Plan: cov.sh

Reviewed By: RaoulFoaleng

Differential Revision: D25894472

Pulled By: ricomariani

fbshipit-source-id: f1f489306b13cc383ac2093200c9a9bd4d250ec2