CQL: Initial cut at discriminated types e.g. integer<task_id>
Summary:
We've long had the ability to detect that object<list> is not assignable to object<dict>; this genrealizes that
notion for all primitive types.  Any primitive type can have a "kind" e.g. integer<task_id> the kind is task_id.
If you then do something like this:

```
declare t integer<task_id> not null;
declare j integer<job_id> not null;

set t := j; -- this is now illegal
```

The rule is that the kind must match in addition to the other conversion rules.  If there is no kind then it mixes freely as usual.  This is how
the `object` type behaves today.

Note that you can currently declare tables with a kind in their columns (SQLite will never see this) but enforcement for table operations like
insert and update is not yet implemented and will require further changes in the semantic type storage.

Reviewed By: RaoulFoaleng

Differential Revision: D25933466

fbshipit-source-id: c53dec37f895339178d22df6a93ea59db2d06a81