Added general support for the notion of unit type "kind" in sem_struct (#41)
Summary:
Used this to enforce kind matches in

 * insert statements
 * compatibility between rows of compound selects
 * argument types of procedure calls
 * fetch cursor from (various flavors)

There are still more places where the enforcement needs to go but the value is now plumbed through
everywhere and so it's just a matter of adding the checks.  Still to come:

  * update
  * compatibility checks for multiple out and out union
  * pretty much everywhere else assignment compatibly is normally checked (easy audit, there's like 6 or so)

Also there are more operators that need to do work in particular
  * unary operators should be sure to preserve kind if it is present
  * cast needs to add or remove it as specified

But even with all this missing we're already catching a ton of invalid cases and adding these
others is very simple indeed.

Landing as is because what we have is strictly better than what we've ever had before and it's
an optional feature anyway.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/41

Reviewed By: Strafos

Differential Revision: D25940554

Pulled By: ricomariani

fbshipit-source-id: 86b8900bae997cf877cf503f8015671d652b0bc6