T76115086 -- window index ranges can only be validated if they are constants or constant expressions
Otherwise we have to check the type only.  Constants are sufficiently common that we use the "eval"
logic to try to evaluate the expression as a constant and report an error if so.  This actually
simplifies the code while adding more funcationalty.  "Eval" did not exist when this check was
first authored.