T76115086 -- window index ranges can only be validated if they are constants or constant expressions (#42)
Summary:
Otherwise we have to check the type only.  Constants are sufficiently common that we use the "eval" logic to try to evaluate the expression as a constant and report an error if so.  This actually simplifies the code while adding more functionality.  "Eval" did not exist when this check was first authored.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/42

Reviewed By: RaoulFoaleng

Differential Revision: D25948332

Pulled By: ricomariani

fbshipit-source-id: 0dbc6d45dccdf5bc2ea54dc21ab1341ba91ff33d