Added "kind" support in the cast operator. You can use it to change or strip kinds.
Summary:
The cast operator changes the kind to be exactly what is in the case, which is either the given kind or no kind.  This can be used to strip change the kind of any expresson.

I refreshed the railroad diagrams at which time I noticed we still had a lex alias for LONG_INT so I fixed that.  This helps the syntax colorer to see everything.

This concludes all known function and expression types for "kind" it's plumbed everywhere now!

Reviewed By: RaoulFoaleng

Differential Revision: D25979220

fbshipit-source-id: 967512270e0730a6d8d94e97f62b28e0d3e716f8