HIDDEN column attribute must be the first one to comply with SQLite needs and CQL needs
Summary:
SQLite treats the hidden attribute as part of the type name but CQL thinks its an attribute like not null.

The happy ground is to insist that it comes first then it looks right to everyone and you can't get it wrong.

Enforce this.  Details in the error doc.

Found a couple of latent warnings, added -Werror to avoid any such issues in the future.

Reviewed By: RaoulFoaleng

Differential Revision: D26140283

fbshipit-source-id: 0272b3d0f4174ffd923e206fda7a16f51f18f2f2