Replace lots of special cases that avoid creating the global proc with general logic
Summary:
If there is any code that runs outside of a proc (e.g. set x:= 1;) then we need a "global proc" to put it in. You provide
that on the command line.  We detect that we need a global proc by looking at the statements generated for it.  If the
buffer is non-empty then we need one.  A classic bug has been that a statement that doesn't need code ends up adding
some because the auto-comment for the statement gets put into the main buffer.

To workaround this we detected "all" the statements that don't generate code and suppressed the comment for them.  This
was fragile and as it turns out still not quite right.

This change replaces the ad hoc tests with logic that simply looks at the code that was generated, if there was no
code put into the main buffer then we do not add the comment there, because (1) that means we would force the global proc
and (2) there's nothing to comment anyway... because nothing was generated.

To do this, we need to see what we're going to generate, so we capture it into temp buffers first, look at it then add
the comment and emit the temp buffer.   Classic CQL trick.

This is also one less land mine for CQL devs to step on because they forgot.

Reviewed By: RaoulFoaleng

Differential Revision: D26382600

fbshipit-source-id: 6379740d871fb6a8bd2b7464b42658814cfbba64