use cql_blob_equal func to compare blobs in codegen for IS, IS NOT, ==, !=, IN and NOT IN operators
Summary:
Previously `==`, `!=`, `IS` operator we doing a pointer comparison when the operand where of type blob. This diff changes that to have blob type behaving a bit like string type where the pointer and the contents of the byte array is compared.  This value-based comparison is consistent with SQLite and probably we should have done it this way in the first place rather than making it like `object`.

Now we can do `blob1 IS blob2` or ``blob1 IS NOT blob2`` or `blob1 == blob2` or `blob1 != blob2`.

Reviewed By: ricomariani

Differential Revision: D26181731

fbshipit-source-id: 501c1f95a049678ec8b8e43d2b7685bccb186db3