Added support for "strict select if nothing"
Summary:
Since (select...) can be problematic (people forget to deal with the case where the select
expression returns no rows) we've added a strict mode to disallow the error-prone version.

This can of course be disabled for cases where you know there will be a row or where you're using
sqlite for its flexible expression evaluation and there is maybe not even a FROM clause.

Reviewed By: RaoulFoaleng

Differential Revision: D26554944

fbshipit-source-id: 4bb3b4a9ebfbd0e7231ef38d9228f229da2301cf