Doc update for @RC
Summary:
We've long needed a way to look at the current error code, especially in the context of
a catch block.  The `RC` builtin variable addresses this allowing you to look at the
thrown error and take various actions depending on what it is.  This is important if
you are trying to handle a particular error.

Reviewed By: wanaya

Differential Revision: D26571742

fbshipit-source-id: e9db55e0ac3adf5c6715c69197e5be29f5e88c1d