When enforcing strict select if nothing, trival selects are ok
Summary:
`enforce_strict select if nothing` is designed to disallow any unsafe uses of
the `(select ...)` form.  In particular it insists that you handle the "no rows" case.

However, there are lots of select expressions were all you want is SQLite evaluation,
such as for getting the current time, or doing a complex conversion,  or some such.
Basically any instance in which there is no `FROM` clause in the select is for sure
safe because it has to return a row.  So e.g.

```
set x := (select 1);
```

Doesn't need to generate an error.  Here we use the sql formatter to detect the presence
of the `from` clause. We do it this way because the select tree is complex and an ad hoc
tree walker would be a bug farm.  The SQL generator has a nice callback facility so we
just added a callback when we hit a from clause and then the standard tree walker does
job for us.

Reviewed By: RaoulFoaleng

Differential Revision: D26594793

fbshipit-source-id: 40495c25831826d85d499e6d1b5b7446de609243