Do not allow the use of "if nothing" forms inside of SQL contexts
Summary:
The (select ... if nothing) construct is for use in top level expressions, not inside of other DML

This form allows for error control of (select...) expressions.  But SQLite does not
understand the form at all, so it can only appear at the top level of expressions where
CQL can strip it out. Here are some examples:

good:

  set x := (select foo from bar where baz if nothing 0);
  if (select foo from bar where baz if nothing 1) then ... end if;

bad:

  select foo from bar where (select something from somewhere if nothing null);
  delete from foo where (select something from somewhere if nothing 1);

Basically if you are already in a SQL context, the form isn't usable because SQLite
simply doesn't understand if nothing at all. This error makes it so that you'll
get a build time failure from CQL rather than a run time failure from SQLite.

Reviewed By: RaoulFoaleng

Differential Revision: D26642372

fbshipit-source-id: de76645b136d58886f83a67084178004966705c6