ifnull_crash is now a synonym for attest_notnull -- symmetric with ifnull_throw
Summary:
For converting to nonnull type once a check has already been made e.g.

if x is not null
  call foo(ifnull_crash(x));  -- foo requires not null arg
end if;

This name better matches ifnull_throw which THROWS if the arg is null otherwise returns a non-null verison.

The attest_notnull version is deprecated and will soon be gone.

Reviewed By: RaoulFoaleng

Differential Revision: D26790503

fbshipit-source-id: 4aa18136d413c58147e7e6897df40505d7bb753b