Add error checking on usage of table valued functions to avoid SQLite "left join to TVF" bug
Summary:
To use TVF safely we have to detect the bad case and report it.  There is an easy workaround described in the help.

This bug was fixed in SQlite 3.31 but we still use lower versions than that.

The new error summarizes the situation nicely:

This error is generated by `enforce_strict table function` it is there to allow safe use of Table Valued Functions (TVF) even though there was a bug in SQLite previous to v 3.31.0 when joining against them.  The bug appears when the TVF is on
the right of a left join. For example:

```
select * from foo left join some_tvf(1);
```

In this case the join becomes an INNER join even though you wrote a left join.  Likewise

```
select * from some_tvf(1) right join foo;
```

Becomes an inner join.  And similarly with cross joins.

The workaround is very simple.  You don't want the TVF to be the target of the join directly.  Instead:

```
with tvf_(*) as (select * from some_tvf(1))
select * from foo left join tvf_;
```

OR

```
select * from foo left join (select * from some_tvf(1));
```

Reviewed By: winniequinn

Differential Revision: D27130444

fbshipit-source-id: 93c85edcd188088c8d321bc88bbded7cb15aaad2