When expanding select * we have to have a good name for all the columns...
Summary:
This is important because if any of the columns are anonymous we have no way to name
them such that SQLite will get it right.

We must expand `*` at *compile* time because there may be deleted columns that must not be included.

So... we can't let SQLite expand it for us... and that means we need a name for everything.

Note:  T.* still has this problem. That one is trickier to check, it will land separately.

Reviewed By: winniequinn

Differential Revision: D27168432

fbshipit-source-id: 9eb306784928ee741e8e800cbc535129e0970555