Disallow comparing against NULL with = and <>
Summary:
There's never a good reason to write `x = NULL` or `x <> NULL`; `x IS NULL` and `X IS NOT NULL` should be used instead.

Specifically   `x = NULL` is NEVER true or false it is ALWAYS `NULL` which is never what you want.  If you want always null, write null...

If you write the above you invariably meant "IS" or "IS NOT"

Reviewed By: ricomariani

Differential Revision: D27164629

fbshipit-source-id: 610c437f8533c7efa832bf64ea0ad8875ba72816