Allow INSERT USING SELECT form
Summary:
The standard form

```
insert into foo(a, b, c, d, e)
 select
     something a,
     something b,
     something c,
     something d,
     something e
     from where etc;
```
Is error prone and redundant especially when the column names are complex.

This diff allows the USING form of INSERT when a SELECT clause is present.

```
insert into foo USING
 select
     something a,
     something b,
     something c,
     something d,
     something e
     from where etc;
```

The column names are implied from the `select`.

Reviewed By: winniequinn

Differential Revision: D27271916

fbshipit-source-id: 4529e8feff493ff80be0f4907336c83ee22f575a