Use of ifnull_throw or throw requires that the procedure returns a result code hence it is a DML proc
Summary:
Even though it's possible to write a procedure that throws that doesn't use the database we
don't want to have many different signatures.  We have vanilla (no failure, no database) and
database which includes the result code.  If you throw you have to be on the 2nd plan.

This diff forces you onto that plan if you throw.

Reviewed By: Strafos

Differential Revision: D27349773

fbshipit-source-id: d057bdb550d2501a3b3050058b4e0480e03b9cf9