One of the call forms was totally not supported; call out_union_proc(); (#56)
Summary:
if you have an existing out union procedure you should be able to call it for your result, which makes you
look like an out union procedure.  This is the same as all the other flavors.

There can be no mix/match here but if you want to multiplex out union procs
you should be able to do this.

The out arg for out union was not properly initialized, it is now.  You'll get a null result for sure rather than garbage.

There are follow up tasks:

* if the call is done in a loop we should release the existing result if there is one (this is done for the cursor cases, that code needs to be moved)
* if an out union proc produces no result (NULL), it should instead produce an empty result (like statements), this can only happen in the call case (the call might not happen because control flow).

In the interest of not making this already tricky diff even bigger doing those separately.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/56

Reviewed By: winniequinn

Differential Revision: D27344546

Pulled By: ricomariani

fbshipit-source-id: e9d7cfc97a31b36a464ab3079c8e16b1719f1177