changing @RC semantics so that they refer to the error in scope
Summary:
The problem here is the initial semantics of `RC` were not well thought out.  The purpose of it is to allow you to see the code that caused the exception path to be taken.  To do this we exposed the internal variable `_rc_` but this turns out to be a big mistake.   The `_rc_` variable has a variety of interim states that might be observable.  This creates lots of cases where looking at it will simply be wrong, you'll see something like `SQLITE_DONE` after a step operation and such.   At first we attempted to hide all those but as this evolved it became clear we had done this wrong.

The code already had a variable `_rc_thrown_` which was initialized in a catch block so that when you do `throw;` the correct exception could be raised (which is really setting the `_rc_` variable back to that value.  This value is actually much closer to what `RC` users really want:  to be able to see what went wrong in scope.

But further analysis revealed that `_rc_thrown_` was itself not quite right.  The problem here is that catch nests inside of other catches and you don't want lose the original exception if an inner catch runs.  Here's an example:

```
begin try
   -- do something
end try;
begin catch
   -- start recovery (
   let original_err = rc;
   begin try
   end try;
   begin catch
      let nested_err = rc;
      -- secondary recovery, suppose this 100% works
   end catch;

   -- additional steps
   if badness then
      throw; -- this should throw original_err!
   end if;
end catch;
```

With a single shared variable `_rc_thrown_` the value of the exception would be lost by the time the throw happened if the inner catch ran.  Any other interior catch in some other nesting scope would likewise lose the error code.

The solution is that each catch block now gets its own saved result variable.  The meaning of `RC` is to give access to that variable, it's lexically scoped so you always get the right error.  At the outer-most level, there is no error, you just get `SQLITE_OK` if you ask.

* `_rc_` goes back to being hidden internal state.
* `_rc_thrown_` is generalized
* semantic analysis doesn't automatically map `rc` into `_rc_thrown_` the codegen pass decides what variable to use.

Specifically:
* each catch block captures the `_rc_` code when it is entered into its own local variable
  * this variable is created lazily, so it only exists if it is used
  * the variable is called `_rc_thrown_n` where n is the catch block number
* any reference to `rc` refers to the capture variable of the innermost catch block the reference is in
* if the `rc` reference happens outside of any catch block its value is `SQLITE_OK`

Reviewed By: winniequinn

Differential Revision: D27404990

fbshipit-source-id: 2de0a032622c4b302e7a435d832ceeee9cd98f37