When making calls to procedures that return a relational result we must free any existing result
Summary:
The simplest example of this is like so:

```
create proc bar()
begin
  select 1 x;
end;

create proc foo()
begin
  call bar();
  call bar(); -- we must finalize our result statement before calling bar or we leak
end;
```

The same thing can happen if calling a procedure that uses `out union`.  In that case
we have to release the result set.

The call and capture in a cursor cases are easier and already handled.

Reviewed By: RaoulFoaleng

Differential Revision: D27415995

fbshipit-source-id: ba89aa523c86b9f68e4243e84ef3df06fa6ddd6c