generalize attribute migrate proc syntax to allow for builtin operations
Summary:
We need a way to have some useful built-in schema migrations.  For now the one we want is one that will let us go from an `recreate` table to an `create` table.  This presently doesn't work right.

This diff allows for migration procs to have a qualified name like `cql:from_recreate`.

These are then checked against the valid set of builtin-migrations and the places they are allowed to appear.  This is exceedingly easy as there is presently only one such migration and it can go in exactly one place.

Notwithstanding, it's been plumbed through generally so that other migrations could be added.  Each one will have to add some complexity to the upgrader so this is not a thing that should be done lightly.

At this point `cql:from_recreate` can flow through in the correct place, and only the correct place, but in the end it does nothing.  The upgrader work is coming in a subsequent diff.

Pending:

* do the actual custom migration
* allow the move recreate to create if and only if this migrator is present

Reviewed By: RaoulFoaleng

Differential Revision: D27488326

fbshipit-source-id: 07e302d5141951a4b23f962e82ef22f0cc94aa28