Temporarily allow changing virtual tables from no migration to cql:module_must_not_be_deleted_see_docs_for_CQL0392
Summary:
This lets us move to a conformant state from a non-conformant state
Enforcement will follow.

Basicaly here we create cql:module_must_not_be_deleted_see_docs_for_CQL0392 but don't require it.  And we allow it to be added to a virtual table delete.  This prepares us for enforcement in the next diff.

Reviewed By: RaoulFoaleng

Differential Revision: D27689594

fbshipit-source-id: a12bbe8ec6f3d74b172d00d5d1d23fc148690a2a