require cql:module_must_not_be_deleted_see_docs_for_CQL0392 on all virtual table deletes
Summary:
To avoid the likely mistake of trying to remove the module for virtual tables we insist that you add this annotation
when you try to delete a virtual table.

It doesn't do anything but it's a reminder that the module should not be deleted and it should go a long way towards
preventing mistakes and avoiding surprises.

Reviewed By: Strafos

Differential Revision: D27686715

fbshipit-source-id: c58ab124229aef340863aa13af4fa2a95d351b6d