Versioning attributes are not allowed on all DDL
Summary:
Versioning attributes are not allowed on DDL that appears in a procedure, that DDL always runs as is

Versioning attributes are not allowed on temp objects (previously only tables were checked), so no `create` on a temp view or anything like that.

If you have a temp view/trigger/table in your previous schema you are allowed to just remove it entirely in future with no tombstone.

Reviewed By: RaoulFoaleng

Differential Revision: D27887854

fbshipit-source-id: 78424f596003f3f2a7cab96e10b211b9dc80f866